+++ client/src/components/Posts/PostForm.jsx
@@ -1,22 +1,22 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { addPost } from '../../redux/post/actions';
+import React, { useState } from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { addPost } from '../../redux/post/actions'
 
 const PostForm = ({ addPost }) => {
-  const [text, setText] = useState('');
+  const [text, setText] = useState('')
 
   return (
-    <div class='post-form'>
-      <div class='bg-primary p'>
+    <div className='post-form'>
+      <div className='bg-primary p'>
         <h3>Say Something...</h3>
       </div>
       <form
-        class='form my-1'
+        className='form my-1'
         onSubmit={e => {
-          e.preventDefault();
-          addPost({ text });
-          setText('');
+          e.preventDefault()
+          addPost({ text })
+          setText('')
         }}
       >
         <textarea
@@ -28,14 +28,14 @@ const PostForm = ({ addPost }) => {
           placeholder='Create a post'
           required
         ></textarea>
-        <input type='submit' class='btn btn-dark my-1' value='Submit' />
+        <input type='submit' className='btn btn-dark my-1' value='Submit' />
       </form>
     </div>
-  );
-};
+  )
+}
 
 PostForm.propTypes = {
   addPost: PropTypes.func.isRequired
-};
+}
 
-export default connect(null, { addPost })(PostForm);
+export default connect(null, { addPost })(PostForm)