+++ client/src/components/Posts/Posts.jsx
@@ -1,17 +1,17 @@
-import React, { useEffect } from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { getPosts } from '../../redux/post/actions';
-import Spinner from '../layout/Spinner';
-import PostItem from './PostItem';
-import PostForm from './PostForm';
+import React, { useEffect } from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { getPosts } from '../../redux/post/actions'
+import Spinner from '../layout/Spinner'
+import PostItem from './PostItem'
+import PostForm from './PostForm'
 
 //
 
 const Posts = ({ getPosts, post: { posts, loading } }) => {
   useEffect(() => {
-    getPosts();
-  }, [getPosts]);
+    getPosts()
+  }, [getPosts])
 
   return loading ? (
     <Spinner />
@@ -21,21 +21,21 @@ const Posts = ({ getPosts, post: { posts, loading } }) => {
       <p className='lead'>
         <i className='fa fa-user'></i> Welcome to the community
       </p>
+
       <PostForm />
-      {posts.map(post => (
-        <PostItem key={post._id} post={post} />
-      ))}
+
+      {posts.length > 0 && posts.map(post => <PostItem key={post._id} post={post} />)}
     </>
-  );
-};
+  )
+}
 
 Posts.propTypes = {
   getPosts: PropTypes.func.isRequired,
   post: PropTypes.object.isRequired
-};
+}
 
 const mapStateToProps = state => ({
   post: state.post
-});
+})
 
-export default connect(mapStateToProps, { getPosts })(Posts);
+export default connect(mapStateToProps, { getPosts })(Posts)