+++ client/src/components/post/Post.jsx
@@ -1,17 +1,17 @@
-import React, { useEffect } from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { getPost } from '../../redux/post/actions';
-import Spinner from '../layout/Spinner';
-import PostItem from '../Posts/PostItem';
-import { Link } from 'react-router-dom';
-import CommentForm from './CommentForm';
-import Comment from './Comment';
+import React, { useEffect } from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { getPost } from '../../redux/post/actions'
+import Spinner from '../layout/Spinner'
+import PostItem from '../Posts/PostItem'
+import { Link } from 'react-router-dom'
+import CommentForm from './CommentForm'
+import Comment from './Comment'
 
 const Post = ({ getPost, post: { post, loading }, match }) => {
   useEffect(() => {
-    getPost(match.params.id);
-  }, getPost);
+    getPost(match.params.id)
+  }, [getPost, match.params.id])
 
   return loading || !post ? (
     <Spinner />
@@ -38,16 +38,16 @@ const Post = ({ getPost, post: { post, loading }, match }) => {
         </div>
       )}
     </>
-  );
-};
+  )
+}
 
 Post.propTypes = {
   getPost: PropTypes.func.isRequired,
   post: PropTypes.object.isRequired
-};
+}
 
 const mapStateToPorps = state => ({
   post: state.post
-});
+})
 
-export default connect(mapStateToPorps, { getPost })(Post);
+export default connect(mapStateToPorps, { getPost })(Post)