+++ client/src/redux/post/actions.js
@@ -1,4 +1,4 @@
-import axios from 'axios';
+import axios from 'axios'
 // import {setAlert} from '../alert/actions'
 import {
   GET_POSTS,
@@ -9,152 +9,148 @@ import {
   GET_POST,
   ADD_COMMENT,
   REMOVE_COMMENT
-} from '../types';
-import { setAlert } from '../alert/actions';
+} from '../types'
+import { setAlert } from '../alert/actions'
 
 export const getPosts = () => async dispatch => {
   try {
-    const res = await axios.get('/api/posts');
+    const res = await axios.get('/api/posts')
 
     dispatch({
       type: GET_POSTS,
       payload: res.data
-    });
+    })
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
 export const getPost = postId => async dispatch => {
   try {
-    const res = await axios.get(`/api/posts/${postId}`);
+    const res = await axios.get(`/api/posts/${postId}`)
 
     dispatch({
       type: GET_POST,
       payload: res.data
-    });
+    })
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
 export const addLike = postId => async dispatch => {
   try {
-    const res = await axios.put(`/api/posts/like/${postId}`);
+    const res = await axios.put(`/api/posts/like/${postId}`)
 
     dispatch({
       type: UPDATE_LIKES,
       payload: { postId, likes: res.data }
-    });
+    })
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
 export const removeLike = postId => async dispatch => {
   try {
-    const res = await axios.put(`/api/posts/unlike/${postId}`);
+    const res = await axios.put(`/api/posts/unlike/${postId}`)
 
     dispatch({
       type: UPDATE_LIKES,
       payload: { postId, likes: res.data }
-    });
+    })
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
-export const deletePost = postId => async dispatch => {
+export const deletePost = id => async dispatch => {
   try {
-    await axios.delete(`/api/posts/${postId}`);
+    await axios.delete(`/api/posts/${id}`)
 
     dispatch({
       type: DELETE_POST,
-      payload: postId
-    });
+      payload: id
+    })
 
-    dispatch(setAlert('Post deleted', 'success'));
+    dispatch(setAlert('Post deleted', 'success'))
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
 export const addPost = formData => async dispatch => {
   const config = {
     headers: {
       'Content-Type': 'application/json'
     }
-  };
+  }
 
   try {
-    const res = await axios.post(`/api/posts`, formData, config);
+    const res = await axios.post(`/api/posts`, formData, config)
 
     dispatch({
       type: ADD_POST,
       payload: res.data
-    });
+    })
 
-    dispatch(setAlert('Post Added', 'success'));
+    dispatch(setAlert('Post Added', 'success'))
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
 export const addComment = (postId, formData) => async dispatch => {
   const config = {
     headers: {
       'Content-Type': 'application/json'
     }
-  };
+  }
   try {
-    const res = await axios.post(
-      `/api/posts/comment/${postId}`,
-      formData,
-      config
-    );
+    const res = await axios.post(`/api/posts/comment/${postId}`, formData, config)
     dispatch({
       type: ADD_COMMENT,
       payload: res.data // all the comments
-    });
-    dispatch(setAlert('Comment Added', 'success'));
+    })
+    dispatch(setAlert('Comment Added', 'success'))
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}
 
 export const removeComment = (postId, commentId) => async dispatch => {
   try {
-    await axios.delete(`/api/posts/${postId}/${commentId}`);
+    await axios.delete(`/api/posts/${postId}/${commentId}`)
     dispatch({
       type: REMOVE_COMMENT,
       payload: commentId
-    });
-    dispatch(setAlert('Comment deleted', 'success'));
+    })
+    dispatch(setAlert('Comment deleted', 'success'))
   } catch (err) {
     dispatch({
       type: POSTS_ERROR,
       payload: { msg: err.response.statusText, status: err.response.status }
-    });
+    })
   }
-};
+}