+++ client/src/redux/post/reducer.js
@@ -7,17 +7,17 @@ import {
   GET_POST,
   ADD_COMMENT,
   REMOVE_COMMENT
-} from '../types';
+} from '../types'
 
 const initialState = {
   posts: [],
   post: null,
   loading: true,
   error: {}
-};
+}
 
 export default (state = initialState, action) => {
-  const { type, payload } = action;
+  const { type, payload } = action
 
   switch (type) {
     case GET_POSTS:
@@ -25,35 +25,35 @@ export default (state = initialState, action) => {
         ...state,
         posts: payload,
         loading: false
-      };
+      }
 
     case GET_POST:
       return {
         ...state,
         post: payload,
         loading: false
-      };
+      }
 
     case ADD_POST:
       return {
         ...state,
         posts: [payload, ...state.posts],
         loading: false
-      };
+      }
 
     case DELETE_POST:
       return {
         ...state,
-        posts: state.posts.map(post => post._id !== payload.id),
+        posts: state.posts.filter(post => post._id !== payload),
         loading: false
-      };
+      }
 
     case POSTS_ERROR:
       return {
         ...state,
         error: payload,
         loading: false
-      };
+      }
 
     case UPDATE_LIKES:
       return {
@@ -62,27 +62,25 @@ export default (state = initialState, action) => {
           post._id === payload.postId ? { ...post, likes: payload.likes } : post
         ),
         loading: false
-      };
+      }
 
     case ADD_COMMENT:
       return {
         ...state,
         post: { ...state.post, comments: payload },
         loading: false
-      };
+      }
 
     case REMOVE_COMMENT:
       return {
         ...state,
         post: {
           ...state.post,
-          comments: state.post.comments.filter(
-            comment => comment._id !== payload
-          )
+          comments: state.post.comments.filter(comment => comment._id !== payload)
         }
-      };
+      }
 
     default:
-      return state;
+      return state
   }
-};
+}