+++ CHANGELOG.md
@@ -1,1 +1,17 @@
+## CLIENT
+
+- check posts before trying to display
+- PostItem : check for item properties before display
+- posts reducer : filter instead of map and payload === id
+- other minor changes
+
+## API
+
+- posts route : fix delete post
+
+---
+
+## TO-DO
+
 - fix the redux bug ? not yet
+- when deleting post error occurs